0.0.3 BUG FIX and UPDATE
Major bug fixed.
JS and CMZ found a tricky bug in the dynamics when Hk(t) is
explicitly Hermitian (i.e. has imaginary off-diagonal terms) as
for instance the BHZ model or the simpler Graphene.

In this case the dynamic was not conserving orbital occupation
during dynamics.

FIX: In the numerical solution of VIE2 and VIDE for multi-orbital
systems the matrix-product matmul(A,G) where A=[1+dt/2*H + Knn*G]^-1
should be replaced by matmul(G,A). Reasoning about the true reason
for this to occurs (it does not follow the calculation I did).

Code under advanced testing.